Create REST API
This is Rest Api the app calls it whenever it needs data from the database. It requests the data from the database based on the parameters sent to the API accordingly certain SQL query is called. Then It converts data into Jason and then sends it as a reply to the application.